pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gspread>=6.2.0",
    "gunicorn>=23.0.0",
    "oauth2client>=4.1.3",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "telethon>=1.39.0",
    "trafilatura>=2.0.0",
]
